Handle CEST/CET Timezones (#2)
This should fix Issue https://github.com/hmatt1/tm-cotd-countdown/issues/1

The plugin works with absolute times. We get the current epoch time, convert to UTC, and then countdown to hard coded cup hours in UTC.

UTC is not affected by daylight savings. Nadeo changes the cup of the day time to follow CEST/CET.

https://www.trackmania.com/2021/08/11/cup-of-the-day-is-now-available-three-times-a-day/

This fix adds logic to determine whether or not we are in CEST/CET using the day of the week calculation as described here: https://artofmemory.com/blog/how-to-calculate-the-day-of-the-week/

Referencing wikipedia as a source, the time switches occur on the last Sunday in March, and the last sunday in October.
https://en.wikipedia.org/wiki/Central_European_Time